增加表单保存相关代码

+++ src/main/java/com/snjtjj/webapi/RptSystemInfoAPI.java
@@ -1,7 +1,10 @@
 package com.snjtjj.webapi;
 
+import com.snjtjj.entity.RptTab;
+import com.snjtjj.entity.RptTabExample;
 import com.snjtjj.entity.SystemInfo;
 import com.snjtjj.entity.SystemInfoExample;
+import com.snjtjj.mapper.RptTabMapper;
 import com.snjtjj.mapper.SystemInfoMapper;
 import com.snjtjj.vo.FormResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,11 +20,13 @@ public class RptSystemInfoAPI {
     @Autowired
     private SystemInfoMapper systemInfoMapper;
 
+    @Autowired
+    private RptTabMapper rptTabMapper;
+
     @PostMapping("/save")
     public FormResponse<String> post(SystemInfo info){
         String id = info.getId();
         FormResponse formResponse;
-        //noinspection Duplicatesext
         if(id != null && (id.toUpperCase().indexOf("KIT") >= 0 || id.toUpperCase().indexOf("EXT") >= 0)){
             info.preInsert();
             systemInfoMapper.insert(info);
@@ -44,7 +49,15 @@ public class RptSystemInfoAPI {
 
     @PostMapping("/delete")
     public FormResponse<String> delete(String id){
+        RptTabExample example = new RptTabExample();
+        example.createCriteria().andSysteminfoidEqualTo(id);
+        List<RptTab> list = rptTabMapper.selectByExample(example);
+        if(list.size() > 0){
+            FormResponse response = new FormResponse("存在明细数据无法删除");
+            response.setSuccess(false);
+            return response;
+        }
         systemInfoMapper.deleteByPrimaryKey(id);
-        return new FormResponse("保存成功！");
+        return new FormResponse("删除成功！");
     }
 }
